pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "google-gen"
version = "3.0.0"
description = "Basic scripts to generate images over the command line using Google's genai APIs"
requires-python = ">=3.8"
authors = []
dependencies = [
  "google-genai>=0.3.0",
  "pillow>=10.0.0"
]

[project.scripts]
google-generate = "google_gen:main"